# server/src/main/java/org/openapitools/api/UsersUser.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.25.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.User;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-09-25T14:47:41.853547+02:00[Europe/Paris]", comments = "Generator version: 7.25.0")
@Validated
@Tag(name = "users", description = "the users API")
public interface UsersUser {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_USER = "/users";
    /**
     * POST /users : Create a user
     *
     * @param user  (required)
     * @return the created user (status code 201)
     *         or the user already exists with the same id (status code 409)
     */
    @Operation(
        operationId = "createUser",
        summary = "Create a user",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "201", description = "the created user", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "409", description = "the user already exists with the same id")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersUser.PATH_CREATE_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> createUser(
        @Parameter(name = "User", description = "", required = true) @Valid @RequestBody User user
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 0, \"name\" : \"name\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_USER_BY_ID = "/users/{id}";
    /**
     * DELETE /users/{id} : Delete a specific user
     *
     * @param id The id of the user to delete (required)
     * @return OK (status code 200)
     *         or user not found (status code 404)
     */
    @Operation(
        operationId = "deleteUserById",
        summary = "Delete a specific user",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Boolean.class))
            }),
            @ApiResponse(responseCode = "404", description = "user not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = UsersUser.PATH_DELETE_USER_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<Boolean> deleteUserById(
        @Parameter(name = "id", description = "The id of the user to delete", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_USERS = "/users";
    /**
     * GET /users : List all users
     *
     * @return an array of users (status code 200)
     */
    @Operation(
        operationId = "listUsers",
        summary = "List all users",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "an array of users", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = User.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersUser.PATH_LIST_USERS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<User>> listUsers(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : 0, \"name\" : \"name\" }, { \"id\" : 0, \"name\" : \"name\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_SHOW_USER_BY_ID = "/users/{id}";
    /**
     * GET /users/{id} : Info for a specific user
     *
     * @param id The id of the user to retrieve (required)
     * @return OK (status code 200)
     *         or user not found (status code 404)
     */
    @Operation(
        operationId = "showUserById",
        summary = "Info for a specific user",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "404", description = "user not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersUser.PATH_SHOW_USER_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<User> showUserById(
        @Parameter(name = "id", description = "The id of the user to retrieve", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 0, \"name\" : \"name\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_USER = "/users/{id}";
    /**
     * PUT /users/{id} : Update a specific user
     * All the properties of the user are replaced, except its id (the id of the path is kept).
     *
     * @param id The id of the user to update (required)
     * @param user  (required)
     * @return the updated user (status code 200)
     *         or user not found (status code 404)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Update a specific user",
        description = "All the properties of the user are replaced, except its id (the id of the path is kept).",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "the updated user", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "404", description = "user not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = UsersUser.PATH_UPDATE_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> updateUser(
        @Parameter(name = "id", description = "The id of the user to update", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "User", description = "", required = true) @Valid @RequestBody User user
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 0, \"name\" : \"name\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}