# server/target/classes/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: team@openapitools.org
  description: Small API to test OpenAPI usage
  license:
    name: All rights reserved
    url: http://apache.org/licenses/LICENSE-2.0.html
  title: OpenAPI user management
  version: 1.0.0
servers:
- url: http://localhost:8080/api
paths:
  /users:
    get:
      operationId: listUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/User"
                type: array
          description: an array of users
      summary: List all users
      tags:
      - users
      x-accepts:
      - application/json
      x-tags:
      - tag: users
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: the created user
        "409":
          description: the user already exists with the same id
      summary: Create a user
      tags:
      - users
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: users
  /users/{id}:
    delete:
      operationId: deleteUserById
      parameters:
      - description: The id of the user to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: OK
        "404":
          description: user not found
      summary: Delete a specific user
      tags:
      - users
      x-accepts:
      - application/json
      x-tags:
      - tag: users
    get:
      operationId: showUserById
      parameters:
      - description: The id of the user to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
        "404":
          description: user not found
      summary: Info for a specific user
      tags:
      - users
      x-accepts:
      - application/json
      x-tags:
      - tag: users
    put:
      description: "All the properties of the user are replaced, except its id (the\
        \ id of the path is kept)."
      operationId: updateUser
      parameters:
      - description: The id of the user to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: the updated user
        "404":
          description: user not found
      summary: Update a specific user
      tags:
      - users
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: users
components:
  schemas:
    User:
      example:
        id: 0
        name: name
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
      type: object